Fix bug causing overflow error message to loop
Add 'exit()' to overflow check. If 'is_too_long_to_save' is True,
print the error message and exit.

This bug was caused by commit c6b4aec, which removed a
feature allowing the user to type 'exit' or 'rerun' at any time.

This feature was holding open the prompt in the case that the number
of months to save exceeded 1800 during the primary calculation.
In this case, the user could either end or restart the program.

The feature removed, nothing held the prompt open, so the number
of months would continue to be incremented while this overflow
condition was checked each time.